# Construct UOps patterns using UPat