… test/test_pattern_matcher.py