Followup on #4791. While doing this, I took the chance to use positional arguments wherever it wasn't too hard, let me know if I should reverse these.

For performance, I noticed checking for `is not None` in a couple places before doing the actual pattern check sped things up non-trivially (not a ton, but noticeable). Now when comparing pre/post-UPat changes including this PR, the performance regression seems fairly minor (though I might've expected an overall speed up). There's a bit of variance between runs, so here's the average over 10 runs for each (all on M1 macbook air):

master:
```
forward 29.587 ms
schedule 9.393 ms
lower 1547.822 ms
```

[master with UPat changes reverted](https://github.com/alecchendev/tinygrad/tree/master-zero-upat):
```
forward 29.907 ms
schedule 9.17 ms
lower 1442.089 ms
```

this branch (full UPat changes complete):
```
forward 29.487 ms
schedule 9.389 ms
lower 1447.634 ms
```