@geohot 2024-06-04
Excellent use of process replay. @Qazalin I'm assuming by the long time CI took that it passed? Why is it so slow, shouldn't it just compare a hash?