@Qazalin 2024-06-04
@geohot yes they passed! https://github.com/tinygrad/tinygrad/actions/runs/9357955542/job/25758850270
I'm pushing to update_benchmark to do the same.