@Qazalin 2024-06-04
process replay runs ASSERT_COMPILE=1 against a cache built up from master, maybe no kernels changed for the limited tests we run on ptx CI? https://github.com/tinygrad/tinygrad/blob/master/.github/workflows/test.yml#L487