@alecchendev 2024-06-04
IIUC it looks like the logic added to match an arg set in #4728 would never match the positive case where the arg is in the set, so the patterns added in that PR never got used. With the latest commit here this is fixed, and now the ptx test is failing