@alecchendev 2024-06-04
> @alecchendev hmm, can you prove this with a separate PR for a regression test? maybe by adapting test_arg_set to the master PatternMatcher

#4827