@geohot 2024-06-05
Congrats, assuming this passes master CI, bounty is yours!

Can pay via PayPal or USDC, e-mail george@tinygrad.org to claim